backend/functions/requirements.txt
functions-framework
google-cloud-storage
Pillow